// src/domain/adapters/soccer/stats-schema.ts
export type SoccerRawStats = {
  started?: boolean;
  substituteAppearance?: boolean;
  minutesPlayed?: number;
  goals?: number;
  assists?: number;
  cleanSheet?: boolean;
  goalsConceded?: number;
  saves?: number;
  penaltySaves?: number;
  yellowCards?: number;
  redCards?: number;
  penaltyMisses?: number;
  ownGoals?: number;
  playerOfTheMatch?: boolean;
};